#52: Fixed native library loading for 6.2.12+ and 7.0.3+ versions (#58)
Fixes native library loading which broke, because of a minor behavior
change in recent JDK versions.

From the UDF application, we modify Java libraries path
`java.library.path` to add additional native libraries at runtime.

In the 6.2.12+ and 7.0.0+ versions of Exasol, the change in the JDK does
not reload the default native library path from `java.library.path` if
it is set to null. This causes `NullPointerException` exception when
loading default libraries such as `net`.

We changed the path modification to append the additional path to the
default libraries paths.

Fixes #52.